+++ environment.yml
@@ -1,0 +1,3 @@
+name: EEG_AUDIO_Anotator
+dependencies :
+
