+++ requirements.txt
@@ -1,42 +1,3 @@
-certifi==2024.2.2
-cffi==1.16.0
-charset-normalizer==3.3.2
-click==8.1.7
-contourpy==1.2.1
-cycler==0.12.1
-decorator==5.1.1
-fonttools==4.52.1
-idna==3.7
-Jinja2==3.1.4
-kiwisolver==1.4.5
-lazy_loader==0.4
-MarkupSafe==2.1.5
-matplotlib==3.9.0
-mne==1.7.0
-numpy==1.26.4
-packaging==24.0
-pandas==2.2.2
-pillow==10.3.0
-platformdirs==4.2.2
-pooch==1.8.1
-pycparser==2.22
-pyparsing==3.1.2
-PyQt5==5.15.9
-pyqt5-plugins==5.15.9.2.3
-PyQt5-Qt5==5.15.2
-PyQt5-sip==12.13.0
-pyqt5-tools==5.15.9.3.3
-pyqtgraph==0.13.7
-python-dateutil==2.9.0.post0
-python-dotenv==1.0.1
-pytz==2024.1
-pyxdf==1.16.6
-qt5-applications==5.15.2.2.3
-qt5-tools==5.15.2.1.3
-requests==2.32.2
-scipy==1.13.1
-six==1.16.0
-soundfile==0.12.1
-tqdm==4.66.4
-tzdata==2024.1
-urllib3==2.2.1
+
+
+#-e .
