+++ run.py
@@ -1,0 +1,12 @@
+from PyQt5.QtWidgets import QApplication
+from src.gui.mainDisplay import MainWindow
+import sys
+
+
+
+
+if __name__ == '__main__':
+    app = QApplication(sys.argv)
+    main_window = MainWindow()
+    main_window.show()
+    sys.exit(app.exec_())
